python_subconverter/requirements.txt
fastapi
uvicorn[standard]
requests
PyYAML
tomli
Jinja2
pydantic